Extract GettingStartedErrorModal
