+++ app/components/GettingStartedModal/GettingStartedWrapper.tsx
@@ -1,24 +1,13 @@
 import React from 'react';
-import StageUtils from '../../utils/stageUtils';
-import { Button, Modal } from '../basic';
+import GettingStartedErrorModal from './GettingStartedErrorModal';
 import GettingStartedModal from './GettingStartedModal';
 import useFetchSchemas from './useFetchSchemas';
-
-const t = StageUtils.getT('gettingStartedModal');
 
 const GettingStartedWrapper = () => {
     const [gettingStartedSchema, cloudSetupSchema, error, clearError] = useFetchSchemas();
 
     if (error) {
-        return (
-            <Modal open>
-                <Modal.Header>{t('errorModal.title')}</Modal.Header>
-                <Modal.Content>{t('errorModal.description')}</Modal.Content>
-                <Modal.Actions>
-                    <Button content="Close" onClick={clearError} />
-                </Modal.Actions>
-            </Modal>
-        );
+        return <GettingStartedErrorModal clearError={clearError} />;
     }
 
     if (gettingStartedSchema && cloudSetupSchema) {
